refactor(dash): Moves model selection callback to models page
