+++ amora/dash/app.py
@@ -1,12 +1,10 @@
 import dash
-from dash import Dash, Input, Output, dcc, html
-from dash.development.base_component import Component
+from dash import Dash, dcc, html
 
 from amora.dash.authentication import add_auth0_login
-from amora.dash.components import model_details, side_bar
+from amora.dash.components import side_bar
 from amora.dash.config import settings
 from amora.dash.css_styles import styles
-from amora.models import amora_model_for_name
 
 dash_app = Dash(
     __name__, external_stylesheets=settings.external_stylesheets, use_pages=True
@@ -37,12 +35,3 @@ dash_app.layout = html.Div(
         ),
     ],
 )
-
-
-@dash_app.callback(
-    Output("model-details", "children"),
-    Input("model-select-dropdown", "value"),
-    prevent_initial_call=True,
-)
-def update_model_details(value: str) -> Component:
-    return model_details.component(model=amora_model_for_name(value))
